+++ elastic_libs.py
@@ -35,6 +35,8 @@ def get_latest_dump(dump_list, **kwargs):
 
     """
 
+    dump_list = list(dump_list)
+
     if dump_list:
         search = max([x[4] for x in dump_list])
 
@@ -57,6 +59,8 @@ def list_dumps(dump_list, **kwargs):
 
     """
 
+    dump_list = list(dump_list)
+
     print("{0:25} {1:15} {2:10} {3:10} {4:10} {5:5} {6:5}"
           .format("Database Dump Name", "Status", "Time", "Number",
                   "Shard Information", "", ""))
@@ -75,9 +79,11 @@ def list_repos2(repo_list, **kwargs):
     Description:  Lists the repositories in the Elasticsearch cluster.
 
     Arguments:
-        (input) repo_list -> List of repositories.
+        (input) repo_list -> Dictionary of repositories.
 
     """
+
+    repo_list = dict(repo_list)
 
     print("{0:30} {1}".format("Repository Name", "Location"))